Implemented http.RunService
This is the *real* core.RunService implementation for executing
Calls. The scheduler-for-pcf API docs don't go into detail, but
it appears that this is a simple authenticated GET request
against the provided URL, so that's what we've done here.

In addition to this, we added logging to some of the API
endpoints and to other various pieces of the system.

Also, we did a bit of refactoring to generally make life easier
and created a docker-compose-based dev/test environment.