Corrected a control flow problem in cf.RunService
Prior to this commit, the private `waitForTask` method could,
theoretically, run forever.

The easiest way to solve this would have been to just insert
some `break` statements in to break out of the loop. Rather
than doing this, though, we chose to refactor the system
a bit to remove the N-level conditional nesting and whatnot.

That wait method is now guaranteed to not continue processing
after a finished Task is returned.