Now using the REAL CF API CLIENT
We've moved the mock for CF API interactions to a more appropriate
seam: the service itself.

That is to say, in order to use the scheduler locall for dev/test
purposes, one must run `jecfapi` and provide its absolute URL to
scheduler via the `CF_ENDPOINT` env var.

This is also true for `jeuaa` and `UAA_ENDPOINT`.

Additionally, we added some quick validation for cron expressions
to prevent the creation of bad schedules for calls/jobs.