Add unit test for ops.DeleteCall
Interestingly, we found a possible segfault scenario that should
never happen in practice (the call passed in via the input being
nil would be bad).

We also added an explicit failure scenario to
mock.CallService.Delete() that returns an error if the call's
name is "sad-face" so as to be able to test a deletion failure.